+++ src/main/kotlin/com/beanz/censusviz/records/SavedQueryDTO.kt
@@ -7,5 +7,5 @@ data class SavedQueryDTO(
         val age: List<Int>,
         val sex: Int,
         val color: Int? = null,
-        val curve: List<Int>? = null
+        val curve: List<Double>? = null
 )
